refactor modernc publisher to take the context of the first message instead of a configuration option
Expanded comments related to publishing context to both implementations.
ZombieZen treats the connection as a file handle, so it does not utilize
the context API. Instead, it can accept a closing channel by
conn.SetInterrupt(ctx.Done()) call.

I decided not to set the interrupt because the same connection is used
to send multiple message batches.